backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
spotipy==2.23.0
pylast==5.2.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
passlib==1.7.4
python-jose==3.3.0
bcrypt==4.1.1
requests==2.31.0
musicbrainzngs==0.7.1

